Support EFS volumes in a task container
There is a new variable `task_volumes`. You can create an EFS volume and pass it down to the ECS module so it attaches the volume in a task container.
The `task_volumes` accepts a list (in case you need more than one volume) of maps with keys `file_system_id` (EFS file system ID) and `container_path` (A directory inside the container where the volume will be mounted).

One more fix is an explisite dependency between the website pod module that creates an ASG and ECS cluster. Terraform wanted to destroy the cluster first while AWS requires ASG to be destroyed first. I achieve the correct order by making the ECS cluster a dependency for the website pod.
```
module "pod" {
  source = "infrahouse/website-pod/aws"
...
  depends_on = [
    aws_ecs_cluster.ecs
  ]
```
Seems working.
