+++ .gitignore
@@ -28,7 +28,7 @@ override.tf.json
 # Include tfplan files to ignore the plan output of command: terraform plan -out=tfplan
 # example: *tfplan*
 .terraform.lock.hcl
-test_data/test_module/terraform.tfvars
+terraform.tfvars
 tf.plan
 .idea
 /docs/_build/
