+++ README.md
@@ -12,9 +12,9 @@ module to create a load balancer, autoscaling group, and update DNS.
 For usage see how the module is used in the using tests in `test_data/test_module`.
 
 ```hcl
-module "test" {
+module "httpd" {
   source  = "infrahouse/ecs/aws"
-  version = "~> 0.8"
+  version = "~> 2.3"
   providers = {
     aws     = aws
     aws.dns = aws
@@ -29,6 +29,44 @@ module "test" {
   zone_id                       = data.aws_route53_zone.cicd.zone_id
   internet_gateway_id           = module.service-network.internet_gateway_id
   task_desired_count       = 1
+}
+```
+
+### Mount EFS volume
+
+The module can attach one or more EFS volumes to a container.
+
+Create the EFS volume with a mount point:
+```hcl
+resource "aws_efs_file_system" "my-volume" {
+  creation_token = "my-volume"
+  tags = {
+    Name = "my-volume"
+  }
+}
+
+resource "aws_efs_mount_target" "my-volume" {
+  for_each       = toset(var.subnet_private_ids)
+  file_system_id = aws_efs_file_system.my-volume.id
+  subnet_id      = each.key
+}
+```
+
+Pass the volumes to the ECS module:
+```hcl
+module "httpd" {
+  source  = "infrahouse/ecs/aws"
+  version = "~> 2.3"
+  providers = {
+    aws     = aws
+    aws.dns = aws
+  }
+...
+  task_volumes = {
+    "my-volume" : {
+      file_system_id : aws_efs_file_system.my-volume.id
+      container_path : "/mnt/"
+    }
 }
 ```
 ## Requirements
@@ -107,10 +145,11 @@ module "test" {
 | <a name="input_service_name"></a> [service\_name](#input\_service\_name) | Service name. | `string` | n/a | yes |
 | <a name="input_ssh_key_name"></a> [ssh\_key\_name](#input\_ssh\_key\_name) | ssh key name installed in ECS host instances. | `string` | n/a | yes |
 | <a name="input_task_desired_count"></a> [task\_desired\_count](#input\_task\_desired\_count) | Number of containers the ECS service will maintain. | `number` | `1` | no |
-| <a name="input_task_environment_variables"></a> [task\_environment\_variables](#input\_task\_environment\_variables) | Environment variables passed down to a task. | <pre>list(object({<br>    name : string<br>    value : string<br>  }))</pre> | `[]` | no |
+| <a name="input_task_environment_variables"></a> [task\_environment\_variables](#input\_task\_environment\_variables) | Environment variables passed down to a task. | <pre>list(<br>    object(<br>      {<br>        name : string<br>        value : string<br>      }<br>    )<br>  )</pre> | `[]` | no |
 | <a name="input_task_max_count"></a> [task\_max\_count](#input\_task\_max\_count) | Highest number of tasks to run | `number` | `10` | no |
 | <a name="input_task_min_count"></a> [task\_min\_count](#input\_task\_min\_count) | Lowest number of tasks to run | `number` | `1` | no |
 | <a name="input_task_role_arn"></a> [task\_role\_arn](#input\_task\_role\_arn) | Task Role ARN. The role will be assumed by a container. | `string` | `null` | no |
+| <a name="input_task_volumes"></a> [task\_volumes](#input\_task\_volumes) | Map name->file\_system\_id of EFS volumes defined in task and available for containers to mount. | <pre>map(<br>    object(<br>      {<br>        file_system_id : string<br>        container_path : string<br>      }<br>    )<br>  )</pre> | `{}` | no |
 | <a name="input_zone_id"></a> [zone\_id](#input\_zone\_id) | Zone where DNS records will be created for the service and certificate validation. | `string` | n/a | yes |
 
 ## Outputs
