+++ test_data/httpd_efs/providers.tf
@@ -1,0 +1,12 @@
+provider "aws" {
+  region = var.region
+  assume_role {
+    role_arn = var.role_arn
+  }
+  default_tags {
+    tags = {
+      "created_by" : "infrahouse/terraform-aws-ecs" # GitHub repository that created a resource
+    }
+
+  }
+}
