+++ test_data/httpd_efs/terraform.tf
@@ -1,0 +1,13 @@
+terraform {
+  //noinspection HILUnresolvedReference
+  required_providers {
+    aws = {
+      source  = "hashicorp/aws"
+      version = "~> 5.11"
+    }
+    cloudinit = {
+      source  = "hashicorp/cloudinit"
+      version = "~> 2.3"
+    }
+  }
+}
