+++ test_data/service-network/main.tf
@@ -1,9 +1,13 @@
 module "service-network" {
   source                = "infrahouse/service-network/aws"
-  version               = "~> 2.0"
-  service_name          = var.service_name
+  version               = "~> 2.3"
+  service_name          = "service-network"
   vpc_cidr_block        = "10.1.0.0/16"
   management_cidr_block = "10.1.0.0/16"
+  # must be enabled for EFS
+  enable_dns_hostnames = true
+  enable_dns_support   = true
+
   subnets = [
     {
       cidr                    = "10.1.0.0/24"
