+++ test_data/service-network/terraform.tf
@@ -1,0 +1,9 @@
+terraform {
+  //noinspection HILUnresolvedReference
+  required_providers {
+    aws = {
+      source  = "hashicorp/aws"
+      version = "~> 5.11"
+    }
+  }
+}
