+++ test_data/service-network/terraform.tfvars
@@ -1,0 +1,3 @@
+
+role_arn = "arn:aws:iam::303467602807:role/ecs-tester"
+region = "us-east-2"
