+++ website-pod.tf
@@ -46,4 +46,7 @@ module "pod" {
     Name : var.service_name
     AmazonECSManaged : true
   }
+  depends_on = [
+    aws_ecs_cluster.ecs
+  ]
 }
